requirements.txt
Flask
opencv-python
pandas
torch
Pillow
transformers
google-generativeai
